Game updated to work with GCC4.  Sweet. :D
git-svn-id: svn://10.0.50.55/puffbomb/trunk@2842 d14347b8-48fd-0310-9e84-97907d434d85

+++ Code/Engine2D/Engine2D.cpp
@@ -432,7 +432,7 @@ void cEngine2D::LoadMap( const std::string MapName )
 		}
 		
 		// Sort static objects in to fronts and backs //
-		for ( int idx = 0; idx < StaticObjectInstance.size(); idx++ ) {
+		for ( size_t idx = 0; idx < StaticObjectInstance.size(); idx++ ) {
 			if ( StaticObjectInstance[ idx ].GetFrontPolygonZ() >= Real::Zero ) {
 				FrontStaticObject.push_back( &StaticObjectInstance[ idx ] );
 			}
